backend/pyproject.toml
[tool.poetry]
name = "chrono-clairvoyant-backend"
version = "0.1.0"
description = "Backend Flask per Chrono Clairvoyant — serve il frontend buildato + API AI + scraping Playwright"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
httpx = "^0.27.2"
playwright = "^1.47.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
